Aufteilung des Trackermanagers bzw. in der Index.js in mehrere Dateien auf Grund von ÜBersichtlichkeit
