+++ index.js
@@ -1,10 +1,11 @@
 /*Allgemeines*/
 const { ApolloServer, gql } = require('apollo-server');
 const express = require('express');
+var router = express.Router();
 /*GraphQL*/
 const graphqlHTTP = require('express-graphql');
 const schema = require('./schema/schema');
-const {graphql} = require('graphql');
+const { graphql } = require('graphql');
 const cors = require('cors');
 /*MongoDB*/
 const mongoose = require('mongoose');
@@ -30,8 +31,8 @@ app.use(bodyParser.urlencoded({ extended: false }))
 app.post('/api/form', (req, res) => {
   console.log(req.body);
   nodemailer.createTestAccount((err, account) => {
-    token= req.body.authToken;
-    link="http://localhost:3000/verify"+"?token="+token;
+    token = req.body.authToken;
+    link = "http://localhost:3000/verify" + "?token=" + token;
     console.log("createTestAccount");
     const htmlEmail = `
       <h3>WebFit Registrierung</h3>
@@ -54,25 +55,25 @@ app.post('/api/form', (req, res) => {
     })
 
     let mailOptions = {
-        from: 'WebFit <app@webfit.app>', // sender address
-        to: req.body.email, // list of receivers
-        replyTo: 'app@webfit.app',
-        subject: 'Welcome to Webfit <3', // Subject line
-        text: req.body.message, // plain text body
-        html: htmlEmail // html body
+      from: 'WebFit <app@webfit.app>', // sender address
+      to: req.body.email, // list of receivers
+      replyTo: 'app@webfit.app',
+      subject: 'Welcome to Webfit <3', // Subject line
+      text: req.body.message, // plain text body
+      html: htmlEmail // html body
     };
 
     // send mail with defined transport object
     transporter.sendMail(mailOptions, (error, info) => {
-        if (error) {
-            return console.log(error);
-        }
-        console.log('Message sent: %s', info.messageId);
-        // Preview only available when sending through an Ethereal account
-        console.log('Preview URL: %s', nodemailer.getTestMessageUrl(info));
+      if (error) {
+        return console.log(error);
+      }
+      console.log('Message sent: %s', info.messageId);
+      // Preview only available when sending through an Ethereal account
+      console.log('Preview URL: %s', nodemailer.getTestMessageUrl(info));
 
-        // Message sent: <b658f8ca-6296-ccf4-8306-87d57a0b4321@example.com>
-        // Preview URL: https://ethereal.email/message/WaQKMgKddxQDoou...
+      // Message sent: <b658f8ca-6296-ccf4-8306-87d57a0b4321@example.com>
+      // Preview URL: https://ethereal.email/message/WaQKMgKddxQDoou...
     });
   });
 });
@@ -89,7 +90,7 @@ app.post('/api/resetPassword', (req, res) => {
   console.log(req.body);
   nodemailer.createTestAccount((err, account) => {
     email = req.body.email;
-    link="http://localhost:3000/newPassword"+"?email="+email;
+    link = "http://localhost:3000/newPassword" + "?email=" + email;
     console.log("createTestAccount");
     const htmlEmail = `
       <h3>Neues Passwort für WebFit!</h3>
@@ -108,24 +109,24 @@ app.post('/api/resetPassword', (req, res) => {
     })
 
     let mailOptions = {
-        from: 'WebFit <app@webfit.app>', // sender address
-        to: req.body.email, // list of receivers
-        replyTo: 'app@webfit.app',
-        subject: 'Neues Passwort für WebFit', // Subject line
-        html: htmlEmail // html body
+      from: 'WebFit <app@webfit.app>', // sender address
+      to: req.body.email, // list of receivers
+      replyTo: 'app@webfit.app',
+      subject: 'Neues Passwort für WebFit', // Subject line
+      html: htmlEmail // html body
     };
 
     // send mail with defined transport object
     transporter.sendMail(mailOptions, (error, info) => {
-        if (error) {
-            return console.log(error);
-        }
-        console.log('Message sent: %s', info.messageId);
-        // Preview only available when sending through an Ethereal account
-        console.log('Preview URL: %s', nodemailer.getTestMessageUrl(info));
+      if (error) {
+        return console.log(error);
+      }
+      console.log('Message sent: %s', info.messageId);
+      // Preview only available when sending through an Ethereal account
+      console.log('Preview URL: %s', nodemailer.getTestMessageUrl(info));
 
-        // Message sent: <b658f8ca-6296-ccf4-8306-87d57a0b4321@example.com>
-        // Preview URL: https://ethereal.email/message/WaQKMgKddxQDoou...
+      // Message sent: <b658f8ca-6296-ccf4-8306-87d57a0b4321@example.com>
+      // Preview URL: https://ethereal.email/message/WaQKMgKddxQDoou...
     });
   });
 });
@@ -133,7 +134,7 @@ app.post('/api/resetPassword', (req, res) => {
 /*Password reset*/
 
 /*GraphQL-Server*/
-app.use('/graphql',graphqlHTTP({
+app.use('/graphql', graphqlHTTP({
   schema,
   graphiql: true
 }));//is fired whenever a graphql request comes in
@@ -163,181 +164,16 @@ app.get('/', (req, res) => {
   res.send('WebFit-Server is online')
 })
 
-app.get('/sync', (req, res) => {
+// Trackermanager
 
-  //Globale Variablen
-  trackerid = req.query.trackerid;
-  user = req.query.user;
+var syncsingle = require('./trackermanager/syncsingle');
+app.get('/sync', syncsingle);
 
-  // Prüfung ob alle Daten beim Request korrekt angegeben wurden
-  if ( trackerid == undefined || user == undefined ){
-    res.send("Error #01 - Request invalid")
-  }else{
-      // Loggin eines neuen Request
-          var newDate = new Date();
-          var date = newDate.getDay() + "." + newDate.getMonth() + "." + newDate.getFullYear() + " / " + newDate.getHours() + ":"+ newDate.getMinutes() + ":"+ newDate.getSeconds()
-          console.log("=============")
-          console.log("Neuer API-Sync-Request: " + date)
-          console.log("Tracker: " + trackerid + " // " + "User: " + user)
+var syncfitbit = require('./trackermanager/sync/fitbit')
+app.get('/sync/fitbit', syncfitbit);
 
-      // Abfrage der userID durch den Tracker von der Datenbank
-      // Anschließend überprüfung ob erhaltene Daten mit angegeneben Daten übereinstimmen
-          function query (str) {
-            return graphql(schema, str);
-          }
-            query(`
-            {
-              tracker(id: "${trackerid}") {
-                id,
-                access_token,
-                token_type,
-                trackerModel{
-                  id,
-                  apiLink,
-                  apiLinkRequest
-                },
-                user{
-                  id
-                },
-                lastSync,
-                user_id
-              }
-            }
-            `).then(data => {
-              console.log(data)
-              if (data['data'].tracker == null){
-                res.send("Error #04 - Not found")
-              }else{
-                //Tracker-Daten kommen an
-                // Nun abgleich mit API-Request Daten
-                var tracker = data['data'].tracker.id // WebFit TrackerID
-                var wearhouse_userid = data['data'].tracker.user_id // FITBIT API
-                var dbuser = data['data'].tracker.user.id // WebFit UserID
-                var token = data['data'].tracker.access_token // Bearer Token
-                var token_type = data['data'].tracker.token_type // Bearer Token
-                var lastSync = data['data'].tracker.lastSync // time in ms
-                var apiLink = data['data'].tracker.trackerModel.apiLink // e.g. api.fitbit.com
-                var apiLinkRequest = data['data'].tracker.trackerModel.apiLinkRequest // e.g. api.fitbit.com
-                if (dbuser === user){
-
-                    // Abgleich der TTL
-                    var currentdate = new Date().getTime()
-                    console.log("-------------")
-                    console.log(currentdate)
-                    console.log("-------------")
-                    var time_diff = currentdate - lastSync
-                    var sync = Math.abs(time_diff)
-                    console.log("Sync-Date: "+ lastSync)
-                    console.log("Current-Date: "+ currentdate)
-                    if (sync > 300000){
-
-                      // Sync ist erlaubt
-                      // hier folgt der Warehouse Request
-
-                      var api_request_link = apiLink+wearhouse_userid+apiLinkRequest
-
-                      axios.get(
-
-                        api_request_link,
-                        {
-                          headers: {
-                            "Authorization": token_type + " " +token
-                          }
-                        }
-
-                        )
-                        .then(function(response){
-                          //console.log(response.data);
-                          // Nun die Daten in die Datenbank schreiben
-                          var array_steps = response.data["activities-steps"];
-                         // var array_steps_length = array_steps.length;
-                         array_steps.forEach(element => {
-                           query(`
-                           mutation{
-                            createSteps(
-                             time: "${element.dateTime}",
-                             value: "${element.value}",
-                             trackerId: "${tracker}"
-                           ){
-                             time
-                           }
-                          }
-                           `).then(done => {
-                            query(`
-                              mutation{
-                                updateTracker(id: "${tracker}", lastSync: "${currentdate}"){id}
-                                 }
-                            `)
-                           })
-                         })
-                         res.send("Success")
-                        })
-                        .catch(function (error) {
-                          res.send("Error #05 - Warehouse-API returned an error <br />" + error)
-                        })
-
-
-                    }else{
-                      res.send("Error #03 - Timeout")
-                    }
-
-
-
-                }else{
-                  res.send("Error #02 - Not allowed")
-                }
-              }
-              })
-            }
-
-          })
-
-app.get('/syncall', (req, res) => {
-
-      
-  //Check is request is authorised
-  var tooken = req.query.tooken
-  if (tooken === "4vtDFA9pjo@aevervj§§röS!2Sfda342346rAFDafkdlfa$"){
-  // Abfrage aller Tracker  
-    console.log("=======================")
-    console.log("Complete Sync Tracker Triggerd")
-    console.log("=======================")
-    function query (str) {
-      return graphql(schema, str);
-    }
-    query(`
-    {
-      allTrackers {
-        id
-        user {
-          id
-        }
-      }
-    }
-
-  `).then(data => {
-    var alltracker = data.data.allTrackers
-    var synctracker = 0;
-    alltracker.forEach(element => {
-      var request = "http://projekt-webfit.de:4009/sync?user="+ element.user.id + "&trackerid=" + element.id
-      axios.get(request).then(function(response){
-        console.log(element.id + " // " +response.data)
-        if (response.data === "Success"){
-          synctracker = synctracker + 1
-        }
-      }
-      )
-    })
-    res.send("Tracker werden gesynct")
-    
-
-  })
-
-  // Falls kein Tooken mit übergeben wird  
-  }else{
-    res.send('Keine Berechtigung')
-  }
-
-})
+// Sync all Tracker
+var syncall = require('./trackermanager/syncall');
+app.get('/syncall', syncall);
